requirements.txt
dagshub == 0.4.2
python-dotenv == 1.0.1
numpy == 1.26.4
pandas == 2.2.3
nltk == 3.9.1
scikit-learn == 1.6.0
scipy == 1.14.1
PyYAML == 6.0.2
lightgbm == 4.5.0
matplotlib == 3.10.0
seaborn == 0.13.2
mlflow == 2.19.0